发送前清除串口缓存避免排长队； 修改传给绘图函数的gForce数据类型。

+++ oym8CHWave/main.cpp
@@ -27,8 +27,8 @@ int main(int argc, char *argv[])
   auto pHub = HubManager::getHubInstance(identifer);
   auto gflistener = make_shared<gfSdkQml>(pHub, std::ref(mainwindow));
 
-  QObject::connect(gflistener.get(), SIGNAL(sendDeviceData(QVector<double>)),
-                   &mainwindow, SLOT(on_drawLine(QVector<double>)));
+  QObject::connect(gflistener.get(), SIGNAL(sendDeviceData(QVector<uint8_t>)),
+                   &mainwindow, SLOT(on_drawLine(QVector<uint8_t>)));
 
   QObject::connect(gflistener.get(), SIGNAL(deviceConnected()),
                    &mainwindow, SLOT(handleDeviceConnected()));
